Feature/44 connect dependency modeler and objective modeler (#75)
* Adds a Reference between objectives of the dependency modeler and the objective modeler
* Adds support for creating, deleting and renaming objectives

---------

+++ app/lib/objectivemodeler/buttonbar/OmButtonBar.js
@@ -5,6 +5,7 @@ import {
 import getDropdown from '../../util/Dropdown';
 import {download, upload} from '../../util/FileUtil';
 import { appendOverlayListeners } from '../../util/HtmlUtil';
+import ObjectiveEvents from "../ObjectiveEvents";
 
 
 export default function OmButtonBar(canvas, eventBus, omModeler) {
@@ -55,7 +56,10 @@ export default function OmButtonBar(canvas, eventBus, omModeler) {
             renameObjectiveInput.value = selectObjectiveComponent.value.name;
             renameObjectiveInput.addEventListener("change", function () {
                 renameObjectiveInput.blur();
-                omModeler.getCurrentObjective().name = renameObjectiveInput.value;
+                eventBus.fire(ObjectiveEvents.OBJECTIVE_RENAMING_REQUESTED, {
+                    objective: selectObjectiveComponent.value.objectiveRef,
+                    name: renameObjectiveInput.value
+                });
                 selectObjectiveComponent.showValue(omModeler.getCurrentObjective());
             });
             renameObjectiveInput.addEventListener("focusout", function () {
@@ -76,9 +80,12 @@ export default function OmButtonBar(canvas, eventBus, omModeler) {
     deleteObjectiveButton.title = 'Delete Current Objective';
     deleteObjectiveButton.addEventListener('click', () => {
         var objectiveToDelete = selectObjectiveComponent.value;
-        omModeler.deleteObjective(objectiveToDelete);
+        var shouldDelete = eventBus.fire(ObjectiveEvents.OBJECTIVE_DELETION_REQUESTED, { objective: objectiveToDelete});
+        if (shouldDelete !== false) {
+            // Deletion was not rejected and not handled somewhere else; should not happen when mediator is involved
+            omModeler.deleteObjective(objectiveToDelete);
+        }
         selectObjectiveComponent.showValue(omModeler.getCurrentObjective());
-        repopulateDropdown();
     });
     buttonBar.appendChild(deleteObjectiveButton);
 
@@ -96,8 +103,9 @@ export default function OmButtonBar(canvas, eventBus, omModeler) {
         selectObjectiveMenu.addCreateElementInput(() => {
             var objectiveName = selectObjectiveMenu.getInputValue();
             if (objectiveName && objectiveName.length > 0) {
-                omModeler.addObjective(objectiveName);
-                repopulateDropdown();
+                eventBus.fire(ObjectiveEvents.OBJECTIVE_CREATION_REQUESTED, {
+                    name: objectiveName
+                });
             }
         });
         deleteObjectiveButton.disabled = objectives.length === 0;
@@ -115,8 +123,9 @@ export default function OmButtonBar(canvas, eventBus, omModeler) {
         selectObjectiveMenu.hide = closeOverlay;
     }
 
-
     eventBus.on('import.render.complete', event => selectObjectiveComponent.showValue(event.rootBoard));
+    eventBus.on([ObjectiveEvents.DEFINITIONS_CHANGED], event => repopulateDropdown());
+
 }
 
 OmButtonBar.$inject = [
